Alvaro branch2

+++ controllers/appointment.js
@@ -1,11 +1,15 @@
 const { getDb } = require('../db/connect');
 const ObjectId = require('mongodb').ObjectId;
+const Api400Error = require('../error-handling/api400Error');
+const Api404Error = require('../error-handling/api404Error');
+const Api500Error = require('../error-handling/api500Error');
+const AsyncWrapper = require('../error-handling/AsyncWrapper');
 
 async function getDbCollection() {
   return getDb().db().collection('appointments');
 }
 
-async function findAll(req, res) {
+const findAll = AsyncWrapper.wrapAsync(async (req, res, next) => {
   const collection = await getDbCollection();
   /*
    #swagger.tags = ['Appointments']
@@ -19,34 +23,45 @@ async function findAll(req, res) {
     res.setHeader('Content-type', 'application/json');
     return res.status(200).json(appointments);
   } catch (error) {
-    console.log(error);
-    return res.status(500).json('Oops! Something went wrong');
+    next(error);
   }
-}
+});
 
-async function findOne(req, res) {
-  const id = new ObjectId(req.params.id);
-  const collection = await getDbCollection();
+const findOne = AsyncWrapper.wrapAsync(async (req, res, next) => {
   /*
    #swagger.tags = ['Appointments']
    #swagger.description = 'Find an appointment by id'
   */
 
   try {
+    if (!ObjectId.isValid(req.params.id)) {
+      throw new Api400Error(
+        'Invalid Appointment ID Format',
+        undefined,
+        `The provided appointment ID ${req.params.id} is not a valid ID format.`
+      );
+    }
+    const id = new ObjectId(req.params.id);
+    const collection = await getDbCollection();
     const result = await collection.find({ _id: id });
     const appointments = await result.toArray();
 
-    if (!appointments.length) return res.status(404).json('Record not found');
+    if (!appointments.length) {
+      throw new Api404Error(
+        'Appointment was not found',
+        undefined,
+        `Error occurred while searching appointment with ID ${req.params.id}.`
+      );
+    }
 
     res.setHeader('Content-type', 'application/json');
     return res.status(200).json(appointments[0]);
   } catch (error) {
-    console.log(error);
-    return res.status(500).json('Oops! Something went wrong');
+    next(error);
   }
-}
+});
 
-async function create(req, res) {
+const create = AsyncWrapper.wrapAsync(async (req, res, next) => {
   /*
    #swagger.tags = ['Appointments']
    #swagger.description = 'Create a new appointment'
@@ -54,56 +69,41 @@ async function create(req, res) {
     in: 'body',
     description: 'Appointment information',
     required: true,
-    "schema": {
-      "type": "object",
-      "properties": {
-        "userId": {
-          "example": "698cd7c051e7d72397d7016a"
-        },
-        "clientId": {
-          "example": "69839a8d2c2aa925aa337a77"
-        },
-        "carId": {
-          "example": "69839a192c2aa925aa337a66"
-        },
-        "date": {
-          "example": "02/12/2026"
-        }
-      }
-    }
+    schema: { $ref: '#/definitions/Appointment' }
+   }
   */
   try {
     const db = getDb().db();
 
     if (req.body.userId) {
       if (!ObjectId.isValid(req.body.userId)) {
-        return res.status(400).json('Invalid User ID format');
+        throw new Api400Error('Invalid User ID format');
       }
       const userId = new ObjectId(req.body.userId);
       const userExists = await db.collection('users').findOne({ _id: userId });
-      if (!userExists) return res.status(404).json('User not found');
+      if (!userExists) throw new Api404Error('User not found');
       req.body.userId = userId;
     }
 
     if (req.body.clientId) {
       if (!ObjectId.isValid(req.body.clientId)) {
-        return res.status(400).json('Invalid Client ID format');
+        throw new Api400Error('Invalid Client ID format');
       }
       const clientId = new ObjectId(req.body.clientId);
       const clientExists = await db
         .collection('clients')
         .findOne({ _id: clientId });
-      if (!clientExists) return res.status(404).json('Client not found');
+      if (!clientExists) throw new Api404Error('Client not found');
       req.body.clientId = clientId;
     }
 
     if (req.body.carId) {
       if (!ObjectId.isValid(req.body.carId)) {
-        return res.status(400).json('Invalid Car ID format');
+        throw new Api400Error('Invalid Car ID format');
       }
       const carId = new ObjectId(req.body.carId);
       const carExists = await db.collection('cars').findOne({ _id: carId });
-      if (!carExists) return res.status(404).json('Car not found');
+      if (!carExists) throw new Api404Error('Car not found');
       req.body.carId = carId;
     }
 
@@ -120,16 +120,19 @@ async function create(req, res) {
 
     if (result.acknowledged) {
       return res.status(201).json({ id: result.insertedId });
-    } else throw new Error('Failed to create new appointment');
+    } else {
+      throw new Api500Error(
+        'Appointment Creation Failed',
+        undefined,
+        'Failed to create a new appointment.'
+      );
+    }
   } catch (error) {
-    console.log(error);
-    return res.status(500).json('Oops! Something went wrong');
+    next(error);
   }
-}
+});
 
-async function update(req, res) {
-  const id = new ObjectId(req.params.id);
-  const collection = await getDbCollection();
+const update = AsyncWrapper.wrapAsync(async (req, res, next) => {
   /*
    #swagger.tags = ['Appointments']
    #swagger.description = 'Update an appointment by id'
@@ -157,37 +160,46 @@ async function update(req, res) {
   */
 
   try {
+    if (!ObjectId.isValid(req.params.id)) {
+      throw new Api400Error(
+        'Invalid Appointment ID Format',
+        undefined,
+        `The provided appointment ID ${req.params.id} is not a valid ID format.`
+      );
+    }
+    const id = new ObjectId(req.params.id);
+    const collection = await getDbCollection();
     const db = getDb().db();
 
     if (req.body.userId) {
       if (!ObjectId.isValid(req.body.userId)) {
-        return res.status(400).json('Invalid User ID format');
+        throw new Api400Error('Invalid User ID format');
       }
       const userId = new ObjectId(req.body.userId);
       const userExists = await db.collection('users').findOne({ _id: userId });
-      if (!userExists) return res.status(404).json('User not found');
+      if (!userExists) throw new Api404Error('User not found');
       req.body.userId = userId;
     }
 
     if (req.body.clientId) {
       if (!ObjectId.isValid(req.body.clientId)) {
-        return res.status(400).json('Invalid Client ID format');
+        throw new Api400Error('Invalid Client ID format');
       }
       const clientId = new ObjectId(req.body.clientId);
       const clientExists = await db
         .collection('clients')
         .findOne({ _id: clientId });
-      if (!clientExists) return res.status(404).json('Client not found');
+      if (!clientExists) throw new Api404Error('Client not found');
       req.body.clientId = clientId;
     }
 
     if (req.body.carId) {
       if (!ObjectId.isValid(req.body.carId)) {
-        return res.status(400).json('Invalid Car ID format');
+        throw new Api400Error('Invalid Car ID format');
       }
       const carId = new ObjectId(req.body.carId);
       const carExists = await db.collection('cars').findOne({ _id: carId });
-      if (!carExists) return res.status(404).json('Car not found');
+      if (!carExists) throw new Api404Error('Car not found');
       req.body.carId = carId;
     }
 
@@ -198,40 +210,57 @@ async function update(req, res) {
     const result = await collection.updateOne({ _id: id }, updateFilter);
 
     if (result.matchedCount === 0) {
-      return res.status(404).json('Record not found');
+      throw new Api404Error(
+        'Appointment was not found',
+        undefined,
+        `Error occurred while searching appointment with ID ${req.params.id}`
+      );
     }
 
     if (result.acknowledged) {
       return res.status(200).json('Appointment was updated successfully');
-    } else throw new Error('Failed to update appointment');
+    } else {
+      throw new Api500Error(
+        'Appointment Update Failed',
+        undefined,
+        `Error occurred while updating appointment with ID ${req.params.id}`
+      );
+    }
   } catch (error) {
-    console.log(error);
-    return res.status(500).json('Oops! Something went wrong');
+    next(error);
   }
-}
+});
 
-async function remove(req, res) {
-  const id = new ObjectId(req.params.id);
-  const collection = await getDbCollection();
+const remove = AsyncWrapper.wrapAsync(async (req, res, next) => {
   /*
    #swagger.tags = ['Appointments']
    #swagger.description = 'Remove an appointment by id'
   */
 
   try {
-    const result = await collection.deleteOne({ _id: id });
-
-    if (result.acknowledged && result.deletedCount === 0) {
-      return res.status(404).json('Record not found');
+    if (!ObjectId.isValid(req.params.id)) {
+      throw new Api400Error(
+        'Invalid Appointment ID Format',
+        undefined,
+        `The provided appointment ID ${req.params.id} is not a valid ID format.`
+      );
     }
+    const id = new ObjectId(req.params.id);
+    const collection = await getDbCollection();
+    const result = await collection.deleteOne({ _id: id });
 
     if (result.deletedCount > 0) {
       return res.status(200).json('Appointment was removed successfully');
-    } else throw new Error('Failed to delete appointment');
+    } else {
+      throw new Api404Error(
+        'Appointment was not found',
+        undefined,
+        `Error occurred while deleting appointment with ID ${req.params.id}`
+      );
+    }
   } catch (error) {
-    console.log(error);
-    return res.status(500).json('Oops! Something went wrong');
+    next(error);
   }
-}
+});
 
 module.exports = { findAll, findOne, create, update, remove };
